Refactor: Model -> Scrapper
